"Removed useState and AIToolCard import from BrandStrip component, and refactored Hero component to add new sections and remove old ones."
